requirements.txt
streamlit==1.41.1
tensorflow==2.19.1
opencv-python==4.12.0.88
pillow==11.3.0
pandas==2.3.1
matplotlib==3.10.5
seaborn==0.13.2
scikit-learn==1.6.1
numpy==2.1.3

